requirements.txt
# Core dependencies
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0.1
resend>=2.0.0
rendercv>=1.14
jinja2>=3.1.4
python-dotenv>=1.0.0
tenacity>=8.2.3

# Data handling
pandas>=2.0.0

# Scraping helpers
fake-useragent>=1.4.0
